Refactor name handling: replace formatter with textutil for name normalization and anonymous mapping across multiple modules
