+++ libs/utils/textutil.py
@@ -3,6 +3,8 @@ libs/utils/textutil.py
 """
 
 import os
+import random
+import re
 from enum import Enum, auto
 from math import ceil, floor
 from typing import TYPE_CHECKING, Any
@@ -24,6 +26,111 @@ class ConversionType(Enum):
     """ひらがなをカタカナに変換"""
     KtoH = auto()
     """カタカナをひらがなに変換"""
+
+
+def name_replace(target: str, add_mark: bool = False, not_replace: bool = False) -> str:
+    """
+    表記ブレ修正(正規化)
+
+    Args:
+        target (str): 対象プレイヤー名
+        add_mark (bool, optional): ゲストマークを付与する. Defaults to False.
+        not_replace (bool, optional): ゲスト置換なし(強制/個人戦) Defaults to False.
+
+            - *True*: ゲストを置換しない
+            - *False*: ゲストを置換する
+
+    Returns:
+        str: 表記ブレ修正後のプレイヤー名
+
+    """
+    chk_pattern = [
+        target,  # 無加工
+        str_conv(target, ConversionType.HtoZ),  # 半角 -> 全角
+        str_conv(target, ConversionType.KtoH),  # カタ -> ひら
+        str_conv(target, ConversionType.HtoK),  # ひら -> カタ
+        honor_remove(target),  # 敬称削除
+        honor_remove(str_conv(target, ConversionType.HtoZ)),  # 敬称削除 + 半角 -> 全角
+        honor_remove(str_conv(target, ConversionType.KtoH)),  # 敬称削除 + カタ -> ひら
+        honor_remove(str_conv(target, ConversionType.HtoK)),  # 敬称削除 + ひら -> カタ
+    ]
+    chk_pattern = sorted(set(chk_pattern), key=chk_pattern.index)  # 順序を維持したまま重複排除
+
+    if g.params.individual or not_replace:
+        for name in chk_pattern:
+            if name in g.cfg.member.lists:  # メンバーリスト
+                return name
+            if name in g.cfg.member.all_lists:  # 別名を含むリスト
+                if ret_name := g.cfg.member.resolve_name(name):
+                    return ret_name
+    else:
+        for team in chk_pattern:
+            if team in g.cfg.team.lists:  # チームリスト
+                return team
+
+    # リストに見つからない場合
+    name = honor_remove(target)
+    if g.params.unregistered_replace and not not_replace:
+        name = g.cfg.member.guest_name
+    if name != g.cfg.member.guest_name and add_mark:
+        name = f"{name}({g.cfg.setting.guest_mark})"
+
+    return name
+
+
+def honor_remove(name: str) -> str:
+    """
+    敬称削除
+
+    Args:
+        name (str): 対象の名前
+
+    Returns:
+        str: 敬称を削除した名前
+
+    """
+    honor = r"(くん|さん|ちゃん|クン|サン|チャン|君)$"
+    if re.match(rf".*{honor}", name):
+        if not re.match(rf".*(っ|ッ|ー){honor}", name):
+            name = re.sub(rf"{honor}", "", name)
+
+    return name
+
+
+def anonymous_mapping(name_list: list[str], initial: int = 0) -> dict[str, str]:
+    """
+    名前リストから変換用辞書を生成
+
+    Args:
+        name_list (list[str]): 名前リスト
+        initial (int, optional): インデックス初期値. Defaults to 0.
+
+    Returns:
+        dict[str, str]: マッピング用辞書
+
+    """
+    ret: dict[str, str] = {}
+
+    if g.params.individual:
+        prefix = "Player"
+        id_list = {x["name"]: x["id"] for x in g.cfg.member.info}
+    else:
+        prefix = "Team"
+        id_list = {x["team"]: x["id"] for x in g.cfg.team.info}
+
+    if len(name_list) == 1:
+        name = name_list[0]
+        if name in id_list:
+            idx = id_list[name]
+        else:
+            idx = int(random.random() * 100 + 100)
+        ret[name] = f"{prefix}_{idx + initial:03d}"
+    else:
+        random.shuffle(name_list)
+        for idx, name in enumerate(name_list):
+            ret[name] = f"{prefix}_{idx + initial:03d}"
+
+    return ret
 
 
 def str_conv(text: str, kind: ConversionType) -> str:
